Move the stage change handler into individual namespace.
