+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Stages/Classic/Stage/StageScreen.cs
@@ -6,7 +6,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Beatmaps;
+using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Stages;
 using osu.Game.Rulesets.Karaoke.Edit.Utils;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Stages.Classic.Stage.Settings;
 using osu.Game.Rulesets.Karaoke.Stages.Classic;
@@ -23,8 +23,8 @@ public partial class StageScreen : ClassicStageScreen, IStageEditorStateProvider
     private readonly Bindable<StageEditorEditCategory> bindableCategory = new();
     private readonly Bindable<StageEditorEditMode> bindableEditMode = new();
 
-    [Cached(typeof(IBeatmapClassicStageChangeHandler))]
-    private readonly BeatmapClassicStageChangeHandler beatmapClassicStageChangeHandler;
+    [Cached(typeof(IClassicStageChangeHandler))]
+    private readonly ClassicStageChangeHandler classicStageChangeHandler;
 
     [Cached(typeof(IStageEditorVerifier))]
     private readonly StageEditorVerifier stageEditorVerifier;
@@ -49,7 +49,7 @@ public partial class StageScreen : ClassicStageScreen, IStageEditorStateProvider
     public StageScreen()
         : base(ClassicStageEditorScreenMode.Stage)
     {
-        AddInternal(beatmapClassicStageChangeHandler = new BeatmapClassicStageChangeHandler());
+        AddInternal(classicStageChangeHandler = new ClassicStageChangeHandler());
         AddInternal(stageEditorVerifier = new StageEditorVerifier());
 
         Child = new GridContainer
